Fork the `SelectionSkeleton` type into internal and external variants.
The `SelectionSkeleton` type uses the `TxOut` type. We eventually want
to replace this with `(output, TokenBundle)` pair, where `output` is a
type parameter representing an abstract output.

The coin selection algorithm doesn't care about the identity or nature
of each output, but the wallet does.
